requirements.txt
Adafruit-Blinka==0.3.0
adafruit-circuitpython-neopixel==3.3.4
Adafruit-GPIO==1.0.3
Adafruit-PureIO==0.2.3
dateparser==0.7.0
Pillow==5.3.0
pytz==2018.7
regex==2018.11.22
requests==2.20.1
rpi-ws281x==4.1.0
RPi.GPIO==0.6.5
spidev==3.2
tzlocal==1.5.1
urllib3==1.24.1
